changed softmax and network to allow more than 10 categories, changed label_t to uint8_t due to read errors

+++ cnn_pure_c/layer.h
@@ -3,7 +3,7 @@
 
 #include "data_structure.h"
 
-typedef int label_t;
+typedef uint8_t label_t;
 typedef double weight_t;
 
 typedef enum {
